fix current location display #18
